Move "legacy" sdui into a package
To be removed completely asap
