+++ feature-odyssey/src/main/kotlin/com/hedvig/android/odyssey/sdui/README.md
@@ -1,0 +1,4 @@
+### SDUI
+
+This package contains the old SDUI components of Odyssey, all to be removed as soon as the native 
+implementation reaches feature parity.
